# backend/docs/swagger.yaml
basePath: /api
host: edis.dshub.com.ng
info:
  contact:
    email: support@lifegate.health
    name: LifeGate Support
  description: AI-powered clinical diagnostic platform. Patients interact with an
    AI assistant, physicians review AI-generated reports, and administrators manage
    the full lifecycle.
  license:
    name: Proprietary
  termsOfService: https://lifegate.health/terms
  title: LifeGate API
  version: "1.0"
paths:
...
  /credits/balance:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  balance:
                    type: integer
                  currency:
                    type: string
                type: object
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Get credit balance
      tags:
      - payments
...
  /payments/bundles:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: array
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: List credit bundles
      tags:
      - payments
  /payments/initiate:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bundle selection
        in: body
        name: body
        required: true
        schema:
          properties:
            bundleId:
              type: string
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  paymentLink:
                    type: string
                  txRef:
                    type: string
                type: object
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Initiate payment
      tags:
      - payments
  /payments/transactions:
    get:
      parameters:
      - description: Max records to return (default 50)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  total:
                    type: integer
                  transactions:
                    type: array
                type: object
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Payment transactions
      tags:
      - payments
  /payments/verify:
    post:
      consumes:
      - application/json
      parameters:
      - description: Transaction reference
        in: body
        name: body
        required: true
        schema:
          properties:
            flwTxId:
              type: string
            txRef:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "402":
          description: Payment Required
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Verify payment
      tags:
      - payments
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"